+++ src/main/services/ProjectService.ts
@@ -13,19 +13,8 @@ import { ScannerPipelineFactory } from '../task/scanner/scannerPipelineFactory/S
 import { ProjectKnowledgeExtractor } from '../modules/projectKnowledge/ProjectKnowledgeExtractor';
 import { ReuseIdentificationTask } from '../task/reuseIdentification/ReuseIdentificationTask';
 import ScannerMode = Scanner.ScannerMode;
-import ScannerType = Scanner.ScannerType;
 
 class ProjectService {
-  private setCryptographyScanType(p: Project): void {
-    if (p.getApiKey()) {
-      const types = p.metadata.getScannerConfig().type;
-      types.push(ScannerType.CRYPTOGRAPHY);
-      const uniqueTypes = new Set(types);
-      p.metadata.getScannerConfig().type = Array.from(uniqueTypes);
-    } else {
-      p.metadata.getScannerConfig().type.filter((t) => t !== ScannerType.CRYPTOGRAPHY);
-    }
-  }
 
   public async close(): Promise<IProject> {
     const p = workspace.getOpenProject();
@@ -33,7 +22,6 @@ class ProjectService {
     await p.close();
     return dto;
   }
-
 
   public async unlock(projectName: string) {
     const db:any = modelProvider.workspace;
@@ -90,10 +78,7 @@ class ProjectService {
 
   public async createProject(projectDTO: INewProject) {
     const p = await this.create(projectDTO);
-    // Add crypto scanner config depending on API Key token
-    this.setCryptographyScanType(p);
     p.save();
-
     await ScannerPipelineFactory.getScannerPipeline(projectDTO.scannerConfig.source).run(p);
   }
 
@@ -112,7 +97,7 @@ class ProjectService {
       const p = await workspace.getProject(new ProjectFilterPath(projectPath));
       p.metadata.getScannerConfig().mode = Scanner.ScannerMode.RESCAN;
       // Add crypto scanner config depending on API Key token
-      this.setCryptographyScanType(p);
+      // this.setCryptographyScanType(p);
       // Save only metadata to avoid overwriting the filetree with an empty new one
       // Using p.save() would overwrite the entire project including the filetree
       p.metadata.save();