+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Upcoming changes ...
+### Added
+- Added configurable pipeline stages in project settings (Dependencies, Vulnerabilities, Cryptography, Search Index)
 
 ## [1.28.0] - 2025-12-22
 ### Added